requirements.txt
beautifulsoup4==4.11.1
bs4requests==2.28.1